+++ CHANGELOG.md
@@ -7,3 +7,4 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 ## [Unreleased]
 ### Added
 - Timeseries plot (from https://plotly.com/python/time-series/#time-series-with-range-selector-buttons)
+- Radar plot for scalars